backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
sentence-transformers
pinecone
pydantic
chunkr-ai
python-multipart